Minor refactoring with PrinterLoader class
