+++ src/main/java/org/spartaglobal/views/App.java
@@ -1,13 +1,13 @@
 package org.spartaglobal.views;
+import org.spartaglobal.controllers.SortManager;
 
 /**
- * Hello world!
- *
+ * The program is called to run from this class
+ * Calls the {@link SortManager} class
  */
 public class App 
 {
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
+    public static void main( String[] args ){
+        SortManager.initialiseUserInputs();
     }
 }
